Move PrefixPosition class into ArgumentTokenizer class
