+++ src/main/java/seedu/address/logic/parser/ArgumentTokenizer.java
@@ -166,4 +166,25 @@ public class ArgumentTokenizer {
         return value.trim();
     }
 
+    /**
+     * Represents a prefix's position in an arguments string.
+     */
+    public static class PrefixPosition {
+        private int startPosition;
+        private final Prefix prefix;
+
+        PrefixPosition(Prefix prefix, int startPosition) {
+            this.prefix = prefix;
+            this.startPosition = startPosition;
+        }
+
+        int getStartPosition() {
+            return startPosition;
+        }
+
+        Prefix getPrefix() {
+            return prefix;
+        }
+    }
+
 }
