--- src/main/java/seedu/address/logic/parser/PrefixPosition.java
@@ -1,22 +1,0 @@
-package seedu.address.logic.parser;
-
-/**
- * Represents a prefix's position in an arguments string.
- */
-public class PrefixPosition {
-    private int startPosition;
-    private final Prefix prefix;
-
-    PrefixPosition(Prefix prefix, int startPosition) {
-        this.prefix = prefix;
-        this.startPosition = startPosition;
-    }
-
-    int getStartPosition() {
-        return startPosition;
-    }
-
-    Prefix getPrefix() {
-        return prefix;
-    }
-}
